refactor: domain 패키지를 만들고 도메인을 이동하라
