+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,3 +1,3 @@
-package blackjack
+package blackjack.domain
 
 data class Card(val suit: Suit, val denomination: Denomination)
