+++ src/main/kotlin/blackjack/domain/CardFactory.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 object CardFactory : CardGenerator {
