+++ src/main/kotlin/blackjack/domain/CardGenerator.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 interface CardGenerator {
     fun generateCards(): List<Card>
