+++ src/main/kotlin/blackjack/domain/Denomination.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 /**
  * 끝수.
