+++ src/main/kotlin/blackjack/domain/Name.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 @JvmInline
 value class Name(val name: String?) {
