+++ src/main/kotlin/blackjack/domain/PlayerStatus.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 enum class PlayerStatus {
     /**
