+++ src/test/kotlin/blackjack/domain/GamerTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
